+++ src/main/java/org/example/stockitbe/store/inventory/StoreInventoryController.java
@@ -3,7 +3,7 @@ package org.example.stockitbe.store.inventory;
 import lombok.RequiredArgsConstructor;
 import org.example.stockitbe.common.model.BaseResponse;
 import org.example.stockitbe.store.inventory.model.StoreInventoryDto;
-import org.example.stockitbe.user.model.AuthUserDetails;
+import org.example.stockitbe.user.model.entity.AuthUserDetails;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
