+++ src/main/java/org/example/stockitbe/user/UserController.java
@@ -12,10 +12,10 @@ import org.example.stockitbe.common.jwt.JwtRefreshRepository;
 import org.example.stockitbe.common.jwt.JwtUtil;
 import org.example.stockitbe.common.model.BaseResponse;
 import org.example.stockitbe.common.model.BaseResponseStatus;
-import org.example.stockitbe.user.model.AuthUserDetails;
-import org.example.stockitbe.user.model.User;
-import org.example.stockitbe.user.model.UserDto;
-import org.example.stockitbe.user.model.UserStatus;
+import org.example.stockitbe.user.model.entity.AuthUserDetails;
+import org.example.stockitbe.user.model.entity.User;
+import org.example.stockitbe.user.model.dto.UserDto;
+import org.example.stockitbe.user.model.entity.UserStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
