+++ src/main/java/org/example/stockitbe/user/model/entity/AuthUserDetails.java
@@ -1,4 +1,4 @@
-package org.example.stockitbe.user.model;
+package org.example.stockitbe.user.model.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
