+++ src/main/java/org/example/stockitbe/user/model/entity/UserStatus.java
@@ -1,4 +1,4 @@
-package org.example.stockitbe.user.model;
+package org.example.stockitbe.user.model.entity;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
